mozjs: make libice, libxt deps explicit
When building from scratch on an x64_64 host for an x86_64 cross, mozjs
do_configure is observed to fail pretty often in do_qa_configure,
because /usr/lib/libICE.so is linked against a test binary. This
primarily comes about because the configure script attempts to detect
X11 by the existence of X11/Intrinsic.h, which is supplied by libxt,
which hasn't yet been built.

To fix, add an explicit libxt dependency, and add a libice dependency
just to make sure the libICE link works.

Signed-off-by: Richard Tollerton <rich.tollerton@ni.com>