florence: re-add recipe
Florence is buggy and unmaintained but it's the only lightweight
alternative to other virtual keyboards like onboard which rdepends on
python or qtvirtualkeyboard which pulls qtbase, qtdeclarative etc.

It was previously removed because it depended on libglade in commit
e4cc6c774b1 ("florence: remove") but libglade is now fixed.