apache2: Use CRYPTO_set_id_callback() when present
mod_ssl uses CRYPTO_THREADID_set_callback() to register one of it's
functions as a handler. CRYPTO_THREADID_set_callback() is write-once,
which is problematic when reloading modules. If mod_ssl is unloaded then
re-loaded to a different address, the previously registered callback is
invalid and cannot be updated using this function.

This change modifies mod_ssl to first uses CRYPTO_set_id_callback()
when present, fall back to CRYPTO_THREADID_set_callback() otherwise.

This is a backport of change 1791849 from svn.apache.org.

Signed-off-by: Haris Okanovic <haris.okanovic@ni.com>