Remove bogus COMMON_LICENSE_DIR usage in LIC_FILES_CHKSUM
Using COMMON_LICENSE_DIR in LIC_FILES_CHKSUM of downloaded
code is never the right thing to do, there must be some
license information and that has to be checked instead.

In several cases the claimed LICENSE was also incorrect.

Signed-off-by: Adrian Bunk <bunk@stusta.de>
Signed-off-by: Khem Raj <raj.khem@gmail.com>

Dropped parts of the commit which did not touch the python*-grpcio*
recipes. When rebasing upon OE upstream, upstream's version of this
commit should always be preferred.

Rebased-by: Alex Stewart <alex.stewart@ni.com>